refactor(gaze): Refactor from_numpy and from_pandas into integration module
