--- src/pymovements/gaze/from_pandas.py
@@ -1,49 +1,0 @@
-# Copyright (c) 2023 The pymovements Project Authors
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-"""Module to create a GazeDataFrame from a numpy array."""
-from __future__ import annotations
-
-import pandas as pd
-import polars as pl
-
-from pymovements.gaze.experiment import Experiment
-from pymovements.gaze.gaze_dataframe import GazeDataFrame
-
-
-def from_pandas(
-        data: pd.DataFrame,
-        experiment: Experiment | None = None,
-) -> GazeDataFrame:
-    """Construct a :py:class:`~pymovements.gaze.gaze_dataframe.GazeDataFrame`
-    from a pandas DataFrame.
-
-    Parameters
-    ----------
-    data:
-        Data represented as a pandas DataFrame.
-    experiment : Experiment
-        The experiment definition.
-
-    Returns
-    -------
-    py:class:`~pymovements.GazeDataFrame`
-    """
-    df = pl.from_pandas(data=data)
-    return GazeDataFrame(data=df, experiment=experiment)
